fix(ios): hold splash until auth resolves to eliminate startup flicker
Three coordinated changes:

1. Splash dismissal moved from main.rs (fires on WASM mount) to an
   Effect in app.rs (fires when auth_loading becomes false). On iOS
   the splash now covers the entire auth resolution window, so users
   go directly from branded splash → login or library with no
   intermediate states visible.

2. Added is_ready() to the JS bridge. The auth polling loop now
   exits early when the bridge is ready but the user isn't signed in,
   instead of always running the full 5-second timeout. On iOS this
   fires on the first tick (PAT check is synchronous).

3. dismiss_splash uses requestAnimationFrame instead of a 600ms hold,
   giving navigation Effects one frame to settle before revealing the
   destination view underneath.
