+++ crates/intrada-web/src/app.rs
@@ -9,6 +9,7 @@ use leptos_router::path;
 use leptos_router::NavigateOptions;
 use send_wrapper::SendWrapper;
 use wasm_bindgen::prelude::*;
+use wasm_bindgen::JsCast;
 
 use intrada_core::{Event, Intrada, SessionEvent, ViewModel};
 
@@ -119,8 +120,13 @@ pub fn App() -> impl IntoView {
                     auth_loading.set(false);
                     return;
                 }
+                // Auth bridge is ready but user isn't signed in — no point
+                // polling for 5 more seconds. On iOS this fires on the first
+                // tick (PAT check is synchronous).
+                if js_bridge::is_ready() {
+                    break;
+                }
             }
-            // After 5 seconds, Clerk has loaded but user is not signed in
             auth_loading.set(false);
         });
     }
@@ -144,6 +150,18 @@ pub fn App() -> impl IntoView {
         js_bridge::add_auth_listener(&closure);
         closure.forget(); // leak intentionally — lives for app lifetime
     }
+
+    // ─── Splash screen dismissal ──────────────────────────────────────
+    // On iOS the splash stays visible until auth resolves and navigation
+    // Effects have settled, so the user goes directly from branded splash
+    // to the final destination (login or library) with no intermediate
+    // flicker. On web the splash element is hidden (display:none) so this
+    // is a no-op.
+    Effect::new(move |_| {
+        if !auth_loading.get() {
+            dismiss_splash();
+        }
+    });
 
     // ─── Crux core + app-level reactive state ─────────────────────────
     // Mounted once at the App level (was previously inside the
@@ -419,6 +437,44 @@ fn AuthLoadingScreen() -> impl IntoView {
             </div>
         </div>
     }
+}
+
+/// Fade out and remove the `#app-splash` overlay. The splash has already been
+/// visible since HTML load (iOS) so we skip the hold and go straight to a
+/// 400ms opacity fade, then remove the element. No-op on web where the splash
+/// is `display: none`.
+fn dismiss_splash() {
+    use leptos::web_sys;
+    let Some(window) = web_sys::window() else {
+        return;
+    };
+    let Some(document) = window.document() else {
+        return;
+    };
+    let Some(el) = document.get_element_by_id("app-splash") else {
+        return;
+    };
+    // One-frame delay so navigation Effects that fire in the same reactive
+    // batch have time to update the DOM before we reveal what's underneath.
+    let raf_cb = Closure::once(move || {
+        let el_remove = el.clone();
+        let _ = el
+            .unchecked_ref::<web_sys::HtmlElement>()
+            .style()
+            .set_property("opacity", "0");
+        let remove_cb = Closure::once(move || {
+            el_remove.remove();
+        });
+        if let Some(w) = web_sys::window() {
+            let _ = w.set_timeout_with_callback_and_timeout_and_arguments_0(
+                remove_cb.as_ref().unchecked_ref(),
+                400,
+            );
+        }
+        remove_cb.forget();
+    });
+    let _ = window.request_animation_frame(raf_cb.as_ref().unchecked_ref());
+    raf_cb.forget();
 }
 
 /// Design catalogue route — shows the component catalogue in debug builds,
